backend/requirements.txt
asgiref==3.8.1
certifi==2025.11.12
dj-database-url==3.0.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
joblib==1.5.1
numpy==2.2.5
pandas==2.3.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
